Fix LoadOrthologTables.psql performance issue.
Rewrote the update statements using MERGE so that we don't have to do aggregate more than once for query.

Also added a new temp index to make transcriptOrthologGroup update faster.

Note: the TranscriptAttributes table was dropped from the update join as the filter for 'protein_coding_gene' is not necesarry since we only have such in the TOG table.